api: update - Some bug fixing and improvements
Fixes a lot of typos and improves the CreateAdminUserCommand. The admin
will also be asked to agree to the terms of service. Improves some other
code lines.

+++ api/database/migrations/2020_03_14_000070_create_addresses_table.php
@@ -30,7 +30,7 @@ class CreateAddressesTable extends Migration
             $table->string('postal');
             $table->unsignedInteger('region_id');
             $table->string('street');
-            $table->unsignedInteger('user_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
             $table->timestamp('created_at', 0)->useCurrent();
             $table->timestamp('updated_at', 0)->useCurrent();
